#1175 Separate analytics generator and hide analytics parameters (#1546)
* #1175 Separate analytics generator and hide analytics parameters

* Remove unused code
fix unit test

* Rollback mandatory field (appId)

* fix the onChanges problem

* #1175 - rebased branch
