+++ ng2-components/ng2-activiti-analytics/src/components/analytics-generator.component.css
@@ -1,0 +1,5 @@
+.chart {display: block; width: 100%;}
+
+.analytics-row__entry {
+    cursor: pointer;
+}
