+++ ng2-components/ng2-activiti-analytics/src/components/analytics-report-parameters.component.spec.ts
@@ -395,10 +395,10 @@ describe('AnalyticsReportParametersComponent', () => {
                 responseText: []
             });
         });
-    });
 
-    it('Should convert a string in number', () => {
-        let numberConvert = component.convertNumber('2');
-        expect(numberConvert).toEqual(2);
+        it('Should convert a string in number', () => {
+            let numberConvert = component.convertNumber('2');
+            expect(numberConvert).toEqual(2);
+        });
     });
 });
