+++ ng2-components/ng2-activiti-analytics/src/components/analytics.component.css
@@ -1,29 +1,1 @@
 .chart {display: block; width: 100%;}
-
-.dropdown-widget {
-    width: 100%;
-}
-
-.dropdown-widget__select {
-    width: 100%;
-}
-
-.dropdown-widget__invalid .dropdown-widget__select {
-    border-color: #d50000;
-}
-
-.dropdown-widget__invalid .dropdown-widget__label {
-    color: #d50000;
-}
-
-.dropdown-widget__invalid .dropdown-widget__label:after {
-    background-color: #d50000;
-}
-
-.dropdown-widget__invalid .mdl-textfield__error {
-    visibility: visible !important;
-}
-
-.analytics-row__entry {
-    cursor: pointer;
-}
