+++ ng2-components/ng2-activiti-analytics/src/i18n/en.json
@@ -5,7 +5,8 @@
       "UNKNOWN-WIDGET-TYPE": "UNKNOWN WIDGET TYPE",
       "FILL-PARAMETER": "Fill in the parameters to generate your report",
       "NO-DATA-FOUND": "No data found",
-      "ZERO-DATA-FOUND": "There are only zero values"
+      "ZERO-DATA-FOUND": "There are only zero values",
+      "SETTING-TITLE": "Change report setting"
     }
   },
   "__KEY_REPORTING": {
