+++ ng2-components/ng2-activiti-analytics/src/i18n/it.json
@@ -5,7 +5,8 @@
             "UNKNOWN-WIDGET-TYPE": "TIPO WIDGET SCONOSCIUTO",
             "FILL-PARAMETER": "Riempi tutti i campi per generare il report",
             "NO-DATA-FOUND": "Nessun valore trovato",
-            "ZERO-DATA-FOUND": "Ci sono solo valori che valgono zero"
+            "ZERO-DATA-FOUND": "Ci sono solo valori che valgono zero",
+            "SETTING-TITLE": "Modifica i parametri del report"
         }
     },
     "__KEY_REPORTING": {
@@ -24,23 +25,23 @@
                 "TYPE-FILTERING": "Include tutti gli steps (Deselezionandolo, rimuoverai gli step come start events, gateways, etc.)?"
             },
             "PROCESS-INSTANCES-OVERVIEW": {
-                "PROCESS-DEFINITION": "Process definition",
+                "PROCESS-DEFINITION": "Definizione del processo",
                 "DATE-RANGE": "Intervallo di Date",
                 "SLOW-PROC-INST-NUMBER": "Quanti process instances lenti vuoi mostrare?"
             },
             "TASK-OVERVIEW": {
-                "PROCESS-DEFINITION": "Process definition",
+                "PROCESS-DEFINITION": "Definizione del processo",
                 "DATE-RANGE": "Intervallo di Date",
                 "DATE-RANGE-INTERVAL": "Aggrega date per"
             },
             "TASK-SLA": {
                 "TASK": "Task",
-                "PROCESS-DEFINITION": "Process definition",
+                "PROCESS-DEFINITION": "Definizione del processo",
                 "DATE-RANGE": "Intervallo di Date",
                 "SLA-DURATION": "Qual' é il tempo che questo task necessita per essere completato per rimanere nella SLA?"
             }
         },
-        "PROCESS-STATUS": "Process stato",
-        "TASK-STATUS": "Task stato"
+        "PROCESS-STATUS": "Stato del processo",
+        "TASK-STATUS": "Stato del task"
     }
 }
